Refactor GridMovemnt
Refactoring GridMovement and behaviors during movement.
