+++ Assets/Scripts/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e858a9a7e3864b359fd52780637a99c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
