+++ Assets/Scripts/Character/Interactor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af094f3f6be984d9d9c6cd214c515c95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
