+++ Assets/Scripts/Character/WallDetector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f90f73a893c1e428f962b141fcd4bb27
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
